Test delete staff member process

+++ src/main/java/lk/ijse/cmjd108/LibMgmt2025/controller/StaffController.java
@@ -36,8 +36,19 @@ public class StaffController {
 
     @DeleteMapping
     public ResponseEntity<Void> deleteStaffMember(@RequestParam ("staffId") String staffId){
-        staffService.deleteStaffMember(staffId);
-        return ResponseEntity.noContent().build();
+        if (staffId == null) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+        try {
+            staffService.deleteStaffMember(staffId);
+            return ResponseEntity.noContent().build();
+        } catch (StaffNotFoundException e) {
+            e.printStackTrace();
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        } catch (Exception e) {
+            e.printStackTrace();
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
 
     @PatchMapping(value = "/{staffId}", consumes = MediaType.APPLICATION_JSON_VALUE)
